# lupi_game_server/openapi.yaml
---
openapi: 3.0.3

servers:
  - url: /v1

info:
  title: Lupi Game API
  description: |
    # Least Unique Positive Integer (Lupi)

    Lupi is a fun game, that has been studied a lot and has some maths
    behind how people behave and what really is optimal in some respect.

    ## The Rules of the Game

    - Lupi is a multiplayer game
    - each player picks a positive integer
    - the lowest unique integer wins

  license:
    name: Unlicense
    url: https://unlicense.org/
  version: 1.0.0

tags:
  - name: game
    description: Playing the game
  - name: stats
    description: Game statistics

paths:
  /rounds:
    get:
      parameters:
        - name: before
          in: query
          description: "(Paging) Return rounds before this round id"
          schema:
            type: integer
          example: 4
        - name: page_size
          in: query
          description: "(Paging) Return at most this many rounds at once"
          schema:
            $ref: "#/components/schemas/page_size"
      tags: [stats]
      description: List completed rounds, most recent first, paged
      x-openapi-router-controller: lupi_game_server.api
      operationId: get_rounds
      responses:
        200:
          description: |
            Limited list of rounds
            - round data (IDs and start date, end date, number of participants)
            - and paging link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/list_of_rounds"

    post:
      tags: [game]
      description: Creates a new game round
      x-openapi-router-controller: lupi_game_server.api
      operationId: create_round
      responses:
        201:
          description: A new game round was created
          content:
            application/json:
              schema:
                type: integer
        409:  # conflict
          description: An active game round already exists

  /votes:
    post:
      tags: [game]
      description: Register a user vote
      x-openapi-router-controller: lupi_game_server.api
      operationId: add_vote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/vote"
      responses:
        200:
          description: Vote accepted
        409:  # conflict
          description: Already voted or the round is already closed
        404:
          description: Round is not available

  /rounds/{round_id}/is_completed:
    parameters:
      - $ref: "#/components/parameters/round_id"
    put:
      tags: [game]
      description: Completes a round, calculates winner
      x-openapi-router-controller: lupi_game_server.api
      operationId: set_round_completed
      requestBody:
        content:
          application/json:
            schema:
              type: boolean
      responses:
        204:
          description: |
            Either of
            - game round was completed
            - no change
        409:  # conflict
          description: |
            When value is false and the round is already completed.

            Reopening a round is not supported because completed
            rounds make information public which would allow one
            to win - it is a possible cheating vector.
        404:
          description: Round is not available

  /rounds/current/id:
    get:
      tags: [game]
      description: Return the current round id
      x-openapi-router-controller: lupi_game_server.api
      operationId: get_current_round_id
      responses:
        200:
          description: Current round id
          content:
            application/json:
              schema:
                type: integer
                example: 12
        404:
          description: Not started a round yet

  /rounds/{round_id}:
    parameters:
      - $ref: "#/components/parameters/round_id"
    get:
      tags: [stats, game]
      description: Get round details
      x-openapi-router-controller: lupi_game_server.api
      operationId: get_round
      responses:
        200:
          description: Details of round
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/round_details"
        404:
          description: Round is not available

components:
  parameters:
    round_id:
      name: round_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/round_id"

  schemas:
    round_id:
      type: integer
      example: 12

    vote:
      type: object
      required:
        - round_id
        - name
        - number
      properties:
        round_id:
          $ref: "#/components/schemas/round_id"
        name:
          type: string
        number:
          $ref: "#/components/schemas/voted_number"
      example:
        round: "1"
        name: "voter"
        number: 1

    voted_number:
      type: integer
      minimum: 1
      example: 2

    round:
      # IDs and start date, end date, players (=number of participants)
      type: object
      required:
        - id
        - start_date
      properties:
        id:
          type: integer
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        players:
          type: integer
          minimum: 0

    round_details:
      type: object
      required:
        - id
        - start_date
        - is_completed
      properties:
        id:
          type: integer
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        players:
          type: integer
          minimum: 0
        is_completed:
          type: boolean
        winner:
          type: string
          example: the one who votes smallest
        vote:
          $ref: "#/components/schemas/voted_number"

    list_of_rounds:
      type: object
      required:
        - data
      properties:
        data:
          type: array
          items:
            allOf:
              - $ref: "#/components/schemas/round"
              - type: object
                required:
                  # make all properties required
                  - id
                  - start_date
                  - end_date
                  - players
            example:
              id: 1
              start_date: 2020-08-01T13:24:50,515819204+00:00
              end_date: 2020-08-02T13:24:50,515819204+00:00
              players: 8
        previous:
          $ref: "#/components/schemas/paging"

    paging:
      type: object
      properties:
        before:
          $ref: "#/components/schemas/round_id"
        page_size:
          $ref: "#/components/schemas/page_size"
      required:
        - before
        - page_size
      example:
        before: 4
        page_size: 25

    page_size:
        type: integer
        minimum: 1
        maximum: 100
        default: 25